изменено:      Documents/circuit.png 	изменено:      kicad/Anemometr/Anemometr.kicad_pcb 	изменено:      kicad/Anemometr/Anemometr.pdf 	изменено:      kicad/Anemometr/Anemometr.sch 	изменено:      kicad/Anemometr/Anemometr.sch-bak 	изменено:      kicad/Anemometr/gerber/Anemometr-job.gbrjob 	изменено:      kicad/Anemometr/gerber/Border.gbr 	изменено:      kicad/Anemometr/gerber/Bottom.gbr 	изменено:      kicad/Anemometr/gerber/MaskBottom.gbr 	изменено:      kicad/Anemometr/gerber/MaskTop.gbr 	изменено:      kicad/Anemometr/gerber/NCData.drl 	изменено:      kicad/Anemometr/gerber/Top.gbr 	изменено:      kicad/Anemometr/gerber/anemometr.zip 	изменено:      kicad/Anemometr/gerber/mk-gerber.sh

+++ kicad/Anemometr/gerber/mk-gerber.sh
@@ -5,7 +5,11 @@ if [ -f anemometr.zip ]; then
 fi
 mv Anemometr-B_Cu.gbr Bottom.gbr
 mv Anemometr-B_Mask.gbr MaskBottom.gbr
-mv Anemometr.drl NCData.drl
+if [ -f Anemometr-PTH.drl ]; then
+	mv Anemometr-PTH.drl NCData.drl
+else
+	mv Anemometr.drl NCData.drl
+fi
 mv Anemometr-F_Cu.gbr Top.gbr
 mv Anemometr-F_Mask.gbr MaskTop.gbr
 mv Anemometr-Edge_Cuts.gbr Border.gbr
